have done: draw line between touch points, save drawing path to bitmap in SD card, can be create by appWidget
well. but the picture can't be load from the SD card, so it can't edit
twice, and if we open a appWidget we had open before, it will be force
shutted down coz' trying to read a pic from the SD card.  In fact i
don't know why it says that the pic can't be decode by bitmapFactory
while the file is exist indeed.